+++ persistence/genesis.go
@@ -164,12 +164,10 @@ func (m *persistenceModule) populateGenesisState(state modules.PersistenceGenesi
 }
 
 // TODO(pocket/issues/149): All of the functions below following a structure similar to `GetAll<Actor>`
-//
 //	can easily be refactored and condensed into a single function using a generic type or a common
-//
-// interface.
+//  interface.
 func (p PostgresContext) GetAllAccounts(height int64) (accs []modules.Account, err error) {
-	ctx, tx, err := p.GetCtxAndTx()
+	ctx, tx, err := p.getCtxAndTx()
 	if err != nil {
 		return nil, err
 	}
@@ -193,7 +191,7 @@ func (p PostgresContext) GetAllAccounts(height int64) (accs []modules.Account, e
 
 // CLEANUP: Consolidate with GetAllAccounts.
 func (p PostgresContext) GetAllPools(height int64) (accs []modules.Account, err error) {
-	ctx, tx, err := p.GetCtxAndTx()
+	ctx, tx, err := p.getCtxAndTx()
 	if err != nil {
 		return nil, err
 	}
@@ -212,7 +210,7 @@ func (p PostgresContext) GetAllPools(height int64) (accs []modules.Account, err 
 }
 
 func (p PostgresContext) GetAllApps(height int64) (apps []modules.Actor, err error) {
-	ctx, tx, err := p.GetCtxAndTx()
+	ctx, tx, err := p.getCtxAndTx()
 	if err != nil {
 		return nil, err
 	}
@@ -223,7 +221,7 @@ func (p PostgresContext) GetAllApps(height int64) (apps []modules.Actor, err err
 	var actors []*types.Actor
 	for rows.Next() {
 		var actor *types.Actor
-		actor, height, err = p.GetActorFromRow(rows)
+		actor, height, err = p.getActorFromRow(rows)
 		if err != nil {
 			return
 		}
@@ -231,7 +229,7 @@ func (p PostgresContext) GetAllApps(height int64) (apps []modules.Actor, err err
 	}
 	rows.Close()
 	for _, actor := range actors {
-		actorWithChains, err := p.GetChainsForActor(ctx, tx, types.ApplicationActor, actor, height)
+		actorWithChains, err := p.getChainsForActor(ctx, tx, types.ApplicationActor, actor, height)
 		if err != nil {
 			return nil, err
 		}
@@ -241,7 +239,7 @@ func (p PostgresContext) GetAllApps(height int64) (apps []modules.Actor, err err
 }
 
 func (p PostgresContext) GetAllValidators(height int64) (vals []modules.Actor, err error) {
-	ctx, tx, err := p.GetCtxAndTx()
+	ctx, tx, err := p.getCtxAndTx()
 	if err != nil {
 		return nil, err
 	}
@@ -252,7 +250,7 @@ func (p PostgresContext) GetAllValidators(height int64) (vals []modules.Actor, e
 	var actors []*types.Actor
 	for rows.Next() {
 		var actor *types.Actor
-		actor, height, err = p.GetActorFromRow(rows)
+		actor, height, err = p.getActorFromRow(rows)
 		if err != nil {
 			return
 		}
@@ -260,7 +258,7 @@ func (p PostgresContext) GetAllValidators(height int64) (vals []modules.Actor, e
 	}
 	rows.Close()
 	for _, actor := range actors {
-		actor, err = p.GetChainsForActor(ctx, tx, types.ApplicationActor, actor, height)
+		actor, err = p.getChainsForActor(ctx, tx, types.ApplicationActor, actor, height)
 		if err != nil {
 			return
 		}
@@ -270,7 +268,7 @@ func (p PostgresContext) GetAllValidators(height int64) (vals []modules.Actor, e
 }
 
 func (p PostgresContext) GetAllServiceNodes(height int64) (sn []modules.Actor, err error) {
-	ctx, tx, err := p.GetCtxAndTx()
+	ctx, tx, err := p.getCtxAndTx()
 	if err != nil {
 		return nil, err
 	}
@@ -281,7 +279,7 @@ func (p PostgresContext) GetAllServiceNodes(height int64) (sn []modules.Actor, e
 	var actors []*types.Actor
 	for rows.Next() {
 		var actor *types.Actor
-		actor, height, err = p.GetActorFromRow(rows)
+		actor, height, err = p.getActorFromRow(rows)
 		if err != nil {
 			return
 		}
@@ -289,7 +287,7 @@ func (p PostgresContext) GetAllServiceNodes(height int64) (sn []modules.Actor, e
 	}
 	rows.Close()
 	for _, actor := range actors {
-		actor, err = p.GetChainsForActor(ctx, tx, types.ServiceNodeActor, actor, height)
+		actor, err = p.getChainsForActor(ctx, tx, types.ServiceNodeActor, actor, height)
 		if err != nil {
 			return
 		}
@@ -299,7 +297,7 @@ func (p PostgresContext) GetAllServiceNodes(height int64) (sn []modules.Actor, e
 }
 
 func (p PostgresContext) GetAllFishermen(height int64) (f []modules.Actor, err error) {
-	ctx, tx, err := p.GetCtxAndTx()
+	ctx, tx, err := p.getCtxAndTx()
 	if err != nil {
 		return nil, err
 	}
@@ -310,7 +308,7 @@ func (p PostgresContext) GetAllFishermen(height int64) (f []modules.Actor, err e
 	var actors []*types.Actor
 	for rows.Next() {
 		var actor *types.Actor
-		actor, height, err = p.GetActorFromRow(rows)
+		actor, height, err = p.getActorFromRow(rows)
 		if err != nil {
 			return
 		}
@@ -318,7 +316,7 @@ func (p PostgresContext) GetAllFishermen(height int64) (f []modules.Actor, err e
 	}
 	rows.Close()
 	for _, actor := range actors {
-		actor, err = p.GetChainsForActor(ctx, tx, types.FishermanActor, actor, height)
+		actor, err = p.getChainsForActor(ctx, tx, types.FishermanActor, actor, height)
 		if err != nil {
 			return
 		}